Added blog.js and blog-live-reload
Also included CSS preamble the blog uses, mostly so it does not get forgotten

Blog.js is a modified version of the code snippet that is responsible for making blogs responsive. I do not particulalry like writing inline code in webflow, since it has a slow refresh cycle time, so I moved it to this project. The other benefit is that shiki would be bundled with blog.js now.
Blog-live-reload.js is a special script that only loads when users want to view blogs in real time.